requirements.txt
accelerate==0.19.0
albumentations==1.3.0
altair==4.2.2
dadaptation==3.1
diffusers[torch]==0.18.2
easygui==0.98.3
einops==0.6.0
fairscale==0.4.13
ftfy==6.1.1
gradio==3.36.1
huggingface-hub==0.15.1
lion-pytorch==0.0.6
lycoris_lora==1.8.0
# for loading Diffusers' SDXL
invisible-watermark==0.2.0
# open clip for SDXL
open-clip-torch==2.20.0
opencv-python
prodigyopt==1.0
pytorch-lightning==1.9.0
rich==13.4.1
safetensors==0.3.1
timm==0.6.12
tk==0.1.0
toml==0.10.2
transformers==4.30.2
voluptuous==0.13.1
wandb==0.15.0
localization==0.1.7
shapely==2.0.1
numexpr==2.8.4
h5py==3.9.0
scikit-image==0.21.0
colorama==0.4.6
pretrainedmodels
deepface
imageio[ffmpeg]
-r Matting/requirements.txt
# for kohya_ss library
-e . # no_verify leave this to specify not checking this a verification stage